# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "상담원 로그인",
        "description": "기본적으로 loginId + extension + password 검증 (bcrypt). supervisor/admin 은 extension 없이도 로그인 가능. 성공 시 access(15m) + refresh(14d) 토큰 쌍 반환.",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh token 으로 새 access 발급",
        "description": "refresh 는 요청마다 회전 (이전 것은 revoke 후 새로 발급). 재사용 공격 방지.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "로그아웃",
        "description": "refresh token 의 revokedAt 을 현재 시각으로 설정. 토큰 없어도 멱등 성공.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "summary": "현재 에이전트의 모든 세션 종료",
        "description": "비밀번호 변경/계정 탈취 의심 시 사용.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/me/session": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "현재 로그인 세션 조회",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/handoff": {
      "post": {
        "operationId": "AuthController_createDesktopHandoff",
        "summary": "Desktop handoff token 생성",
        "description": "현재 로그인한 웹 세션이 60초짜리 1회용 handoff token 을 발급해 Windows 런타임이 교환할 수 있게 한다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHandoffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/handoff/exchange": {
      "post": {
        "operationId": "AuthController_exchangeDesktopHandoff",
        "summary": "Desktop handoff token 교환",
        "description": "1회용 handoff token 을 access/refresh token 으로 교환한다. 성공하면 같은 token 은 재사용할 수 없다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeHandoffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/agent-updates/session": {
      "post": {
        "operationId": "AgentUpdatesController_createSession",
        "summary": "업데이트 세션 발급",
        "description": "일반 CTI access token 을 updateSessionToken 으로 교환한다. updateSessionToken 은 manifest 와 download-init 호출에만 사용한다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "agent-updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/agent-updates/manifest": {
      "get": {
        "operationId": "AgentUpdatesController_getManifest",
        "summary": "승인된 데스크톱 앱 manifest 조회",
        "description": "콜센터 서버에 승인된 최신 에이전트 데스크톱 버전과 강제 업데이트 정책을 반환한다. Authorization 헤더에는 updateSessionToken 을 보낸다.",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentVersion",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "agent-updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/agent-updates/download-init": {
      "post": {
        "operationId": "AgentUpdatesController_downloadInit",
        "summary": "업데이트 다운로드 토큰 발급",
        "description": "updateSessionToken 으로 tenant 범위의 승인된 artifact 를 확인한 뒤, artifactId 에 묶인 1회성 downloadToken 을 발급한다.",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadInitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "agent-updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/agent-updates/artifacts/{artifactId}": {
      "get": {
        "operationId": "AgentUpdatesController_downloadArtifact",
        "summary": "업데이트 설치 파일 다운로드",
        "description": "Authorization 헤더의 downloadToken 이 가리키는 artifact 와 요청 artifactId 가 일치할 때만 파일을 내려준다.",
        "parameters": [
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "agent-updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/agent-updates/report": {
      "post": {
        "operationId": "AgentUpdatesController_report",
        "summary": "업데이트 결과 보고",
        "description": "다운로드/설치/롤백 상태와 임의 metadata 를 감사 로그로 누적한다. 이 엔드포인트는 일반 CTI access token 을 사용한다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "agent-updates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "KAster CTI API",
    "description": "Asterisk CTI middleware API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "error": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "loginId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          }
        },
        "required": [
          "loginId",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CreateHandoffDto": {
        "type": "object",
        "properties": {
          "deviceName": {
            "type": "string",
            "example": "front-desk-pc-01"
          }
        }
      },
      "ExchangeHandoffDto": {
        "type": "object",
        "properties": {
          "handoffToken": {
            "type": "string"
          }
        },
        "required": [
          "handoffToken"
        ]
      },
...
      "CreateUpdateSessionDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "example": "pc-001"
          },
          "currentVersion": {
            "type": "string",
            "example": "1.3.2"
          }
        }
      },
      "DownloadInitDto": {
        "type": "object",
        "properties": {
          "artifactId": {
            "type": "string",
            "example": "agent-win-x64-1.4.0"
          },
          "currentVersion": {
            "type": "string",
            "example": "1.3.2"
          }
        },
        "required": [
          "artifactId",
          "currentVersion"
        ]
      },
      "ReportUpdateDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}